server/requirements.txt
alembic==1.7.5
black==21.11b1
cachelib==0.5.0
click==8.0.3
flake8==4.0.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.2
regex==2021.11.10
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.0
Werkzeug==2.0.2
